requirements.txt
# Chronoscope - Programme ORPHÉE
# Python dependencies

# Core
numpy>=1.24.0
Pillow>=10.0.0

# Web interface
gradio>=4.0.0

# Audio processing (optional, for soundscape features)
# pydub>=0.25.0
# simpleaudio>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
